% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_con_survival.R
\name{est_con_survival}
\alias{est_con_survival}
\title{Predict conditional survival probability}
\usage{
est_con_survival(
  L,
  t_hor,
  predict.id,
  predict.id.one,
  new.fu_measure = NULL,
  new.fu_time_variable = NULL,
  new.baseline_value_lmm = NULL,
  new.z_value = NULL,
  new.x_value = NULL,
  AUC = FALSE,
  Brier = FALSE,
  object,
  no_cores = 7
)
}
\arguments{
\item{L}{a value of the landmark time in the same unit as the observed survival times and the original form of measurement times
in the data set.}

\item{t_hor}{a value of the time horizon of prediction in the same unit as the parameter L.
The summation of L and t_hor needs to be less than the maximum of observed survival times in the data set.}

\item{predict.id}{a character for the type of prediction. It could be "all","one", or "new".
predict.id="all" is to do prediction for all patients who are still at risk at time \eqn{L} in the data set.
predict.id="one" is to do prediction for a specific patient who is still at risk at landmark time \eqn{L} in the data set, where the
patient id is specified in the parameter predict.id.one.
predict.id="new" is to do prediction for a new patient who is not in the data set and at risk at landmark time \eqn{L}. This new
patient's new values are specified in the parameters new.fu_measure, new.fu_time_variable, new.baseline_value_lmm, new.z_value and new.x_value.}

\item{predict.id.one}{a patient's id number from the data set when predict.id="one",
which could be numeric and character depending on the form of patient id column in the data set. By default predict.id.one = NULL.}

\item{new.fu_measure}{a vector of the new patient's biomarker measurements up to time \eqn{L} when predict.id="new". All measurements are
in the same form specified for the parameter fu_measure in the fitted longitudinal submodel of jmfhc_point_est().
By default new.fu_measure = NULL.}

\item{new.fu_time_variable}{a vector of the new patient's measurement time points in the forms specified for the parameter
fu_time_variable in the fitted longitudinal submodel of jmfhc_point_est() when predict.id="new". When fu_time_variable is more than one variable,
the elements in new.fu_time_variable is placed in the order of variables in the fu_time_variable.
By default new.fu_time_variable = NULL.}

\item{new.baseline_value_lmm}{value(s) of the baseline covariate(s) in the order specified for the parameter baseline_var_lmm in the fitted
longitudinal submodel of jmfhc_point_est() when predict.id="new". By default new.baseline_value_lmm = NULL.}

\item{new.z_value}{value(s) of the long-term baseline covariate(s) in the order specified for the parameter beta_variable in the fitted
cure submodel of jmfhc_point_est() when predict.id="new". By default new.z_value = NULL.}

\item{new.x_value}{value(s) of the short-term baseline covariate(s) in the order specified for the parameter gamma_variable in the fitted
cure submodel of jmfhc_point_est() when predict.id="new". By default new.x_value = NULL.}

\item{AUC}{a logical value (True or False) to determine whether to evaluate the AUC. By default AUC=FALSE.}

\item{Brier}{a logical value (True or False) to determine whether to evaluate the Brier score. By default Brier=FALSE.}

\item{object}{an object of jmfhc_point_est() function.}

\item{no_cores}{the number of cores used during the estimation. The default is 7.}
}
\value{
a list containing the prediction results depending on the arguments for the parameters predict.id, AUC, and Brier.
The following items are returned.
\item{con_surv} a data.frame with a column of patient id and a column of predicted conditional probabilities of not experiencing the event of
interest in an additional time \eqn{t_{hor}} at time \eqn{L}. When predict.id="all", the data.frame contains predicted probabilities for
all patients who are still at risk at time \eqn{L}. When predict.id="one", the data.frame contains the specified patient's predicted probability.
When predict.id="new", the data.frame contains this new patient's predicted probability.
\item{AUC} the computed time-independent AUC value if AUC=TRUE.
\item{Brier_score} the computed Brier score if Brier=TRUE.
\item{Landmark} the value of landmark time specified in the function argument L.
\item{horizon} the value of time horizon of prediction specified in the function argument t_hor.
}
\description{
Predicts an individual's conditional probability of not experiencing the event of
interest in an additional time \eqn{t_{hor}} given that the patient remains risk-free at least until
time \eqn{L} and evaluates the fitted joint cure model's predictive performance by the time-dependent
area under the curve (AUC) of the receiver operating characteristic (ROC) and Brier score.
}
\examples{
result_jmfhc <- jmfhc::jmfhc_point_est(data=longdat,
                               event_time="event.time", event_status="event",
                               id="patient.id",beta_variable="trt", gamma_variable="trt",
                               fu_measure="measure", fu_time_original="mes.times",
                               fu_time_variable="mes.times")
predict_surv <- est_con_survival(L=10,t_hor=5,predict.id="all",AUC=TRUE,Brier=TRUE,object=result_jmfhc)
predict_surv <- est_con_survival(L=10,t_hor=5,predict.id="one",predict.id.one=3,AUC=FALSE,Brier=FALSE,object=result_jmfhc)
predict_surv <- est_con_survival(L=10,t_hor=5,predict.id="new",
                      new.fu_measure=c(5.2,5.1,1.6,0.9,-1.1,-2.6,-5.3,-8.0,-7.5,-11,-12),
                      new.fu_time_variable= 0:10,
                      new.baseline_value_lmm=NULL,
                      new.z_value=0,
                      new.x_value=0,
                      AUC=FALSE,Brier=FALSE,
                      object=result_jmfhc)
predict_surv

}
