% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_con_surv.R
\name{plot_con_surv}
\alias{plot_con_surv}
\title{Plot a patient's observed biomarker values and predicted conditional survival function}
\usage{
plot_con_surv(
  landmark,
  predict.id,
  model.type,
  object,
  original_biomarker_form,
  no_cores = 7
)
}
\arguments{
\item{landmark}{a value of the landmark time included in the object assigned below.}

\item{predict.id}{a subject id who is still risk free at landmark time in the test dataset.}

\item{model.type}{a character for the type of model used for prediction. It could be "joint model" or "landmark model".}

\item{object}{an object of pred_surv_joint_model() function for JMFHC model or pred_landmark_model() function for the LFHC model.}

\item{original_biomarker_form}{a logical value (True or False) to determine whether to use the originial form of biomarker values biomarker values
presented at the y-axis of the plot.}

\item{no_cores}{the number of cores used during the estimation. The default is 7.}
}
\value{
a plot containing a patient's observed biomarker measurments up to the time \eqn{L} and the predicted conditional survival probability at \eqn{L}
}
\description{
Plots a patient's observed biomarker values up to a landmark time \eqn{L} and predicted
individual conditional survival function at \eqn{L}
}
\examples{
# Example of JMFHC
predict_surv_jmfhc  <- pred_surv_joint_model(landmark_surv_vec=c(10,20),
                                             thor_list=list(c(5,10),c(5,10)),
                                             object=jmfhc_est,
                                             test_dat=test_dat,
                                             no_cores=7)
plot_con_surv(landmark=10,predict.id=1,model.type="joint model",object=predict_surv_jmfhc,original_biomarker_form=TRUE,no_cores=7)

# Example of the LFHC model
lfhc_pred_all <- pred_landmark_model(cure_prob=T,landmark_cure_vec=c(10,20),
                                     surv_prob=T,landmark_surv_vec=c(10,20),
                                     thor_list=list(c(5,10),c(5,10)),
                                     train_prep_data=train_dat,test_prep_data=test_dat,
                                     id="id",event_time="event.time",event_status = "event",
                                     fu_time_original="measure.time",
                                     fu_measure_original="measure",fu_measure="measure",
                                     beta_variable = c("x1","x2"),
                                     gamma_include_measure=FALSE,
                                     gamma_variable= c("x1","x2"))
plot_con_surv(landmark=10,predict.id=1,model.type="landmark model",object=lfhc_pred_all,original_biomarker_form=TRUE,no_cores=7)

}
