% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_cure_joint_model.R
\name{pred_cure_joint_model}
\alias{pred_cure_joint_model}
\title{Predict conditional probabilities of being cured using a joint model}
\usage{
pred_cure_joint_model(landmark_cure_vec, object, test_dat, no_cores = 7)
}
\arguments{
\item{landmark_cure_vec}{a vector of landmark times for prediction of cure probabilities in the same unit as the observed survival times and the original form of measurement times
in the data set.}

\item{object}{an object of jmfhc_point_est() function estimated by JMFHC or JMPHC for the training data.}

\item{test_dat}{a data frame containing observations for prediction}

\item{no_cores}{the number of cores used during the prediction. The default is 7.
#' @return a list containing sublists for each assigned landmark time from landmark_cure_vec.
Within each sublist at each assigned landmark time, there are data frames for time horizons of predictions containing id and
corresponding predicted conditional cure probabilities.}
}
\value{
a list with predicted conditional probabilities at assigned landmark times (landmark_cure_vec).
}
\description{
Predicts an individual's the conditional probability of being cured at a landmark time \eqn{L}
using a joint model with a flexible-hazards cure model for survival data (JMFHC) or
a joint model with a proportional hazards cure model for survival data (JMPHC)
}
\examples{
jmfhc_estresult <- readRDS("jmfhc_estresult.rds")
jmfhc_est <- jmfhc::jmfhc_point_est(data=train_dat, event_time="event.time", event_status="event",
                                    id="id", beta_variable=c("x1","x2"), gamma_variable=c("x1","x2"),
                                    fu_measure_original="measure",fu_measure="measure",
                                    fu_time_original="measure.time",fu_time_fixed_variable="measure.time",
                                    fu_time_random_variable="measure.time",
                                    baseline_var_lmm=c("x1","x2"),no_cores=7)
predict_cure_jmfhc  <- pred_cure_joint_model(landmark_cure_vec =c(10,20),
                                             object=jmfhc_estresult,
                                             test_dat=test_dat,
                                             no_cores=7)
predict_cure_jmfhc$`landmark=10`
predict_cure_jmfhc$`landmark=20`

}
