% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_landmark_model.R
\name{pred_landmark_model}
\alias{pred_landmark_model}
\title{Predict conditional cure or survival probability using a landmark model}
\usage{
pred_landmark_model(
  cure_prob,
  landmark_cure_vec = NULL,
  surv_prob,
  landmark_surv_vec = NULL,
  thor_list = NULL,
  train_prep_data,
  test_prep_data,
  id,
  event_time,
  event_status,
  fu_time_original,
  fu_measure_original,
  fu_measure,
  beta_variable,
  gamma_include_measure = FALSE,
  gamma_variable = NULL,
  no_cores = 7
)
}
\arguments{
\item{cure_prob}{a logical value (True or False) to determine whether to predict conditional cure probabilities for subject(s)
in the data set for prediction (test_prep_data).}

\item{landmark_cure_vec}{a vector of landmark times for prediction of cure probabilities in the same unit as the observed survival times and the original form of measurement times
in the data sets. By default landmark_cure_vec = NULL.}

\item{surv_prob}{a logical value (True or False) to determine whether to predict conditional survival probabilities for subject(s)
in the data set for prediction (test_prep_data).}

\item{landmark_surv_vec}{a vector of landmark times for prediction of survival probabilities in the same unit as the observed survival times and the original form of measurement times
in the data sets. By default landmark_surv_vec = NULL.}

\item{thor_list}{a list of vectors of the time horizons of prediction in the same unit as the parameter landmark_surv_vec in the order of the corresponding landmark times.}

\item{train_prep_data}{a data frame used as a training data for estimation.}

\item{test_prep_data}{a data frame containing subject(s) for prediction.}

\item{id}{the variable name corresponding to id in the train_prep_data and test_prep_data.}

\item{event_time}{the variable name corresponding to the event time in the train_prep_data and test_prep_data.}

\item{event_status}{the variable name corresponding to the event status in the train_prep_data and test_prep_data.}

\item{fu_time_original}{the variable name corresponding to the original form of measurement time points in the train_prep_data and test_prep_data.
Variables event_time and fu_time_original are in the same time unit.}

\item{fu_measure_original}{the variable name corresponding to the longitudinal biomarker measurements in the original form.}

\item{fu_measure}{the variable name corresponding to the longitudinal biomarker measurements in the original form or any transformation used in the model.}

\item{beta_variable}{the names of the long-term covariates besides the baseline and most current biomarker values.}

\item{gamma_include_measure}{a logical value (True or False) to determine whether to include the baseline and most current biomarker values as short-term covariates.
By default gamma_include_measure = FALSE.}

\item{gamma_variable}{the names of the short-term covariates besides the baseline and most current biomarker values. By default gamma_variable = NULL.}

\item{no_cores}{the number of cores used during the prediction. The default is 7.}
}
\value{
a list containing sublists for all assigned landmark times from landmark_cure_vec if cure_prob=TRUE and landmark_surv_vec if surv_prob=TRUE.
Each sublist at each assigned landmark time contains the following items.
\item{train_est} the estimation result of the landmark model containing coefficients (coef), the number of iterations (iter), updated data with
estimated baseline hazard function and baseline cumulative hazard function (data), and a list containing all the specified parameters in the argument for the fitted landmark model (setting).
\item{test_data_landmark} a data frame with risk-free subjects at landmark time.
\item{pred_cure} returns if cure_prob=TRUE. A data frame contains id and its corresponding predicted conditional cure probability.
\item{pred_surv} returns if surv_prob=TRUE. Data frames for time horizons of predictions contain id and
corresponding predicted conditional survival probabilities.
}
\description{
Predicts an individual's conditional probability of being cured at a landmark time \eqn{L} and/or
conditional probability of not experiencing the event of interest in an additional time \eqn{t_{hor}}
given that the patient remains risk-free at least until time \eqn{L} using
a landmark flexible-hazards cure (LFHC) model or a landmark proportional hazards cure (LPHC) model
incorporating the longitudinal biomarker values by including baseline and the most current biomarker values as covariates.
}
\examples{
lfhc_pred_all <- pred_landmark_model(cure_prob=T,landmark_cure_vec=c(10,20),
                                     surv_prob=T,landmark_surv_vec=c(10,20),
                                     thor_list=list(c(5,10),c(5,10)),
                                     train_prep_data=train_dat,test_prep_data=test_dat,
                                     id="id",event_time="event.time",event_status = "event",
                                     fu_time_original="measure.time",
                                     fu_measure_original="measure",fu_measure="measure",
                                     beta_variable = c("x1","x2"),
                                     gamma_include_measure=FALSE,
                                     gamma_variable= c("x1","x2"))
# predicted conditional cure probabilities for the test data at landmark 10 and 20
pred_cure_10 <- lfhc_pred_all$`landmark=10`$pred_cure
pred_cure_20 <- lfhc_pred_all$`landmark=20`$pred_cure

# predicted conditional survival probabilities for the test data
# landmark 10 with thor=5 and 10
pred_surv_L10_thor5 <- lfhc_pred_all$`landmark=10`$pred_surv$`thor=5`
pred_surv_L10_thor10 <- lfhc_pred_all$`landmark=10`$pred_surv$`thor=10`
# landmark 20 with thor=5 and 10
pred_surv_L20_thor5 <- lfhc_pred_all$`landmark=20`$pred_surv$`thor=5`
pred_surv_L20_thor10 <- lfhc_pred_all$`landmark=20`$pred_surv$`thor=10`

}
