% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv_joint_model.R
\name{pred_surv_joint_model}
\alias{pred_surv_joint_model}
\title{Predict conditional survival probability using a joint model}
\usage{
pred_surv_joint_model(
  landmark_surv_vec,
  thor_list,
  object,
  test_dat,
  no_cores = 7
)
}
\arguments{
\item{landmark_surv_vec}{a vector of landmark times for prediction of survival probabilities in the same unit as the observed survival times and the original form of measurement times
in the data set.}

\item{thor_list}{a list of vectors of the time horizons of prediction in the same unit as the parameter landmark_surv_vec in the order of the corresponding landmark times.}

\item{object}{an object of jmfhc_point_est() function estimated by JMFHC or JMPHC for the training data.}

\item{test_dat}{a data frame containing observations for prediction}

\item{no_cores}{the number of cores used during the prediction. The default is 7.}
}
\value{
a list containing sublists for each assigned landmark time from landmark_surv_vec and the object of jmfhc_point_est().
Within each sublist at each assigned landmark time, there are data frames for time horizons of predictions containing id and
corresponding predicted conditional survival probabilities.
}
\description{
Predicts an individual's conditional probability of not experiencing the event of
interest in an additional time \eqn{t_{hor}} given that the patient remains risk-free at least until
time \eqn{L} using a joint model with a flexible-hazards cure model for survival data (JMFHC) or
a joint model with a proportional hazards cure model for survival data (JMPHC)
}
\examples{
jmfhc_est <- jmfhc::jmfhc_point_est(data=train_dat, event_time="event.time", event_status="event",
                                    id="id", beta_variable=c("x1","x2"), gamma_variable=c("x1","x2"),
                                    fu_measure_original="measure",fu_measure="measure",
                                    fu_time_original="measure.time",fu_time_fixed_variable="measure.time",
                                    fu_time_random_variable="measure.time",
                                    baseline_var_lmm=c("x1","x2"),no_cores=7)
# jmfhc_est <- readRDS("jmfhc_estresult.rds")
predict_surv_jmfhc  <- pred_surv_joint_model(landmark_surv_vec=c(10,20),
                                             thor_list=list(c(5,10),c(5,10)),
                                             object=jmfhc_est,
                                             test_dat=test_dat,
                                             no_cores=7)
predict_surv_jmfhc$`landmark=10`$`thor=10`
predict_surv_jmfhc$`landmark=20`$`thor=10`

}
