% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDynShiny.R
\name{runDynPred}
\alias{runDynPred}
\title{Shiny Web App for indiviudal dynamic prediction}
\usage{
runDynPred()
}
\value{
predicted values
}
\description{
Shiny Web App for indiviudal dynamic prediction
}
\examples{
runDynPred()
# For demenstration, jmfhc_estresult.rds is an example file for RDS.
# demo_new_patient.csv is a new patient's records up to the landmark time=10

}
